requirements.txt
discord.py[voice]==1.3.1
youtube_dl
aiohttp[speedups]
async-timeout==3.0.1